Refactored UI components for better modularity and consistency.
- Extracted reusable components including `GameBoard`, `GameCell`, `GameOverDialog`, and `CylinderActionButton` into dedicated files in the `ui.components` package.
- Introduced new common components: `NeonCard`, `NeonBadge`, `SquareIconButton`, and `SettingsComponents`.
- Updated `GameCell` to support a dynamic `fontSizeMultiplier` and improved glassmorphism styling.
- Standardized UI across `GameScreen`, `MainScreen`, `ProfileScreen`, and `ThemeSettingsScreen` using the new shared components.
- Added a `Context.findActivity()` extension to safely handle window insets and status bar colors, especially in Compose Previews.
- Refined `GridSizeBottomSheet` and `GameModeCard` to utilize the new `NeonCard` component.
