+++ app/src/main/java/com/avfusionapps/game_2048/ui/screens/GameScreen.kt
@@ -78,6 +78,8 @@ import com.avfusionapps.game_2048.ui.components.ClassicBottomBar
 import com.avfusionapps.game_2048.ui.components.GameScoreBoard
 import com.avfusionapps.game_2048.ui.components.GameSwipeIndicator
 import com.avfusionapps.game_2048.ui.components.GridSizeBottomSheet
+import com.avfusionapps.game_2048.ui.components.GameBoard
+import com.avfusionapps.game_2048.ui.components.GameOverDialog
 import com.avfusionapps.game_2048.ui.theme.HighLighter
 import com.avfusionapps.game_2048.ui.theme.LocalGameTheme
 import com.avfusionapps.game_2048.ui.theme._2048OriginalTheme
@@ -333,256 +335,4 @@ fun GameScreen(
     }
 }
 
-@Composable
-fun GameBoard(viewModel: GameViewModel) {
-    val theme = LocalGameTheme.current
-    Box(
-        modifier = Modifier
-            .aspectRatio(1f)
-            .fillMaxWidth()
-            .clip(RoundedCornerShape(12.dp))
-            .background(theme.surfaceColor)
-            .border(2.dp, theme.primaryColor.copy(alpha = 0.5f), RoundedCornerShape(12.dp))
-            .padding(6.dp)
-    ) {
-        Column(
-            modifier = Modifier.fillMaxSize(),
-            verticalArrangement = Arrangement.spacedBy(4.dp)
-        ) {
-            viewModel.gameState.grid.forEachIndexed { i, row ->
-                Row(
-                    modifier = Modifier
-                        .weight(1f)
-                        .fillMaxWidth(),
-                    horizontalArrangement = Arrangement.spacedBy(4.dp)
-                ) {
-                    row.forEachIndexed { j, cellValue ->
-                        val tileInfo = viewModel.gameState.tileAnimationInfo[Pair(i, j)]
-                        GameCell(
-                            value = cellValue,
-                            modifier = Modifier
-                                .weight(1f)
-                                .testTag("GameScreen_Item_GameCell_${i}_${j}")
-                                .semantics {
-                                    contentDescription = if (cellValue == 0) "Empty Tile at row $i column $j" else "Tile $cellValue at row $i column $j"
-                                },
-                            isNew = tileInfo?.isNew ?: false,
-                            isMerged = tileInfo?.isMerged ?: false
-                        )
-                    }
-                }
-            }
-        }
-    }
 
-    LaunchedEffect(viewModel.gameState.moveCount) {
-        if (viewModel.gameState.tileAnimationInfo.isNotEmpty()) {
-            delay(300) // Adjust delay based on longest animation
-            viewModel.clearAnimationInfo()
-        }
-    }
-}
-
-@Composable
-fun GameCell(
-    value: Int,
-    modifier: Modifier = Modifier,
-    isNew: Boolean = false,
-    isMerged: Boolean = false
-) {
-    val theme = LocalGameTheme.current
-    val backgroundColor = theme.tileColors[value] ?: theme.tileColors[0]!!
-    val textColor = if (value <= 4) theme.textColor else Color.White
-
-    val textSize = remember(value) {
-        when {
-            value >= 10000 -> 14.sp
-            value >= 1000 -> 20.sp
-            value >= 100 -> 24.sp
-            else -> 28.sp
-        }
-    }
-
-    val cellShadowModifier = if (value in 2..32) {
-        Modifier.shadow(
-            elevation = 12.dp,
-            shape = RoundedCornerShape(6.dp),
-            clip = false,
-            ambientColor = Color.White.copy(alpha = 0.25f),
-            spotColor = Color.White.copy(alpha = 0.25f)
-        )
-    } else {
-        Modifier
-    }
-
-    val textShadow = remember(value) {
-        if (value > 0) {
-            val logValue = log2(value.toFloat()).coerceAtLeast(1f)
-            val baseBlur = 2f
-            val scaleFactor = 1.5f
-            val blurRadius = (baseBlur + (logValue * scaleFactor)).coerceIn(2f, 25f)
-            Shadow(
-                color = if (value > 128) textColor.copy(alpha = 0.2f) else Color.White.copy(alpha = 0.6f),
-                offset = Offset.Zero,
-                blurRadius = blurRadius
-            )
-        } else {
-            null
-        }
-    }
-
-    val borderModifier = if (value in 2..32) {
-        Modifier.border(
-            width = 1.dp,
-            color = Color.White.copy(alpha = 0.3f),
-            shape = RoundedCornerShape(6.dp)
-        )
-    } else {
-        Modifier
-    }
-
-    var scale by remember { mutableStateOf(1f) }
-    LaunchedEffect(key1 = value, key2 = isNew, key3 = isMerged) {
-        scale = 1f // Reset scale
-        if (isNew) {
-            scale = 0.1f
-            animate(0.1f, 1f, animationSpec = tween(150, 50)) { v, _ -> scale = v }
-        } else if (isMerged) {
-            animate(1f, 1.2f, animationSpec = tween(80)) { v, _ -> scale = v }
-            animate(1.2f, 1f, animationSpec = tween(80)) { v, _ -> scale = v }
-        }
-    }
-
-    Box(
-        modifier = modifier
-            .aspectRatio(1f)
-            .graphicsLayer(scaleX = scale, scaleY = scale)
-            .then(cellShadowModifier) // Apply glow
-            .clip(RoundedCornerShape(6.dp))
-            .background(backgroundColor)
-            .then(borderModifier), // Apply border
-        contentAlignment = Alignment.Center
-    ) {
-        if (value != 0) {
-            Text(
-                text = value.toString(),
-                color = textColor,
-                fontSize = textSize,
-                fontWeight = FontWeight.Bold,
-                textAlign = TextAlign.Center,
-                style = TextStyle(
-                    shadow = textShadow
-                ),
-                modifier = Modifier.drawBehind {
-                    val paint = Paint()
-                        .asFrameworkPaint().apply {
-                            this.style = android.graphics.Paint.Style.STROKE
-                            this.strokeWidth = 2f  // Increased stroke width
-                            this.color = Color.White.copy(alpha = 0.7f).hashCode()
-                            this.textSize = textSize.toPx()
-                            this.isAntiAlias = true  // Enable anti-aliasing
-                            this.flags = this.flags or android.graphics.Paint.SUBPIXEL_TEXT_FLAG
-                        }
-
-                    val textBounds = android.graphics.Rect()
-                    val text = value.toString()
-                    paint.getTextBounds(text, 0, text.length, textBounds)
-
-                    drawIntoCanvas { canvas ->
-                        val x = (size.width - textBounds.width()) / 2f
-                        val y = (size.height + textBounds.height()) / 2f
-                        canvas.nativeCanvas.drawText(
-                            text,
-                            x,
-                            y,
-                            paint
-                        )
-                    }
-                }
-            )
-        }
-    }
-}
-
-@Composable
-fun GameOverDialog(score: Int, onNewGame: () -> Unit, onExit: () -> Unit) {
-    val theme = LocalGameTheme.current
-    Dialog(onDismissRequest = {}) {
-        Surface(
-            modifier = Modifier
-                .testTag("GameScreen_Dialog_GameOver")
-                .padding(horizontal = 16.dp, vertical = 24.dp),
-            shape = RoundedCornerShape(16.dp),
-            color = theme.surfaceColor,
-            shadowElevation = 8.dp
-        ) {
-            Column(
-                modifier = Modifier.padding(24.dp),
-                horizontalAlignment = Alignment.CenterHorizontally,
-                verticalArrangement = Arrangement.spacedBy(16.dp)
-            ) {
-                Text(
-                    "Game Over!",
-                    style = MaterialTheme.typography.headlineMedium,
-                    color = HighLighter
-                )
-                Text(
-                    "Final Score: $score",
-                    style = MaterialTheme.typography.titleLarge,
-                    color = Color.White
-                )
-                Spacer(modifier = Modifier.height(8.dp))
-                Row(Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.SpaceEvenly) {
-                    NeonRoundedButton(
-                        modifier = Modifier.testTag("GameScreen_Button_GameOverNewGame"),
-                        onClick = onNewGame,
-                        text = "New Game"
-                    )
-                    NeonRoundedButton(
-                        modifier = Modifier.testTag("GameScreen_Button_GameOverExit"),
-                        onClick = onExit,
-                        text = "Exit"
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Preview
-@Composable
-fun GameCellBoardPreview() {
-    val values = listOf(0, 2, 4, 8, 16, 32, 64, 128, 256, 512, 1024, 2048)
-
-    _2048OriginalTheme {
-        LazyVerticalGrid(
-            columns = GridCells.Fixed(4),
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(16.dp),
-            contentPadding = PaddingValues(16.dp)
-        ) {
-            items(values) { value ->
-                Box(
-                    modifier = Modifier.padding(4.dp)
-                ) {
-                    GameCell(value, Modifier.size(75.dp))
-                }
-            }
-        }
-    }
-}
-
-@Preview
-@Composable
-fun GameCellPreview() {
-    val values = listOf(0, 2, 4, 8, 16, 32, 64, 128, 256, 512, 1024, 2048)
-
-    _2048OriginalTheme {
-        Box(
-            modifier = Modifier.padding(4.dp)
-        ) {
-            GameCell(values[0], Modifier.size(75.dp))
-        }
-    }
-}
